refactor(error): migrate ServerConnect to unified ServerConnectError
